+++ miniconf/src/json_core.rs
@@ -78,7 +78,7 @@ impl<'de, T: TreeSerialize<Y> + TreeDeserialize<'de, Y>, const Y: usize> JsonCor
     ) -> Result<usize, Error<de::Error>> {
         let mut de = de::Deserializer::new(data);
         self.deserialize_by_key(keys.into_keys(), &mut de)?;
-        de.end().map_err(Error::PostDeserialization)
+        de.end().map_err(Error::Finalization)
     }
 
     #[inline]